# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Profile {
  degree?: string;
  branch?: string;
  year?: string;
  university?: string;
  skills: string[];
  targetRole: string;
  targetCompany?: string;
  careerGoal: string;
  experienceLevel: string;
  existingProjects?: string;
  internshipExperience?: string;
  learningAreas?: string;
}

export interface Preferences {
  category: string;
  difficulty: string;
  /**
     * @minimum 1
     * @maximum 8
     */
  count: number;
  preferredTechnologies?: string[];
}

/**
 * @nullable
 */
export type GenerateInputVariationProject = { [key: string]: unknown } | null;

export interface GenerateInput {
  profile: Profile;
  preferences: Preferences;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  variationProject?: GenerateInputVariationProject;
  /** @nullable */
  variationRequest?: string | null;
}

export interface TechStack {
  frontend: string[];
  backend: string[];
  database: string[];
  ai: string[];
  other: string[];
}

export interface Project {
  id: string;
  title: string;
  oneLinePitch: string;
  problem: string;
  solution: string;
  targetUsers: string[];
  category: string;
  difficulty: string;
  resumeValue: number;
  technicalDepth: number;
  uniqueness: number;
  placementRelevance: number;
  techStack: TechStack;
  coreFeatures: string[];
  advancedFeatures: string[];
  architecture: string;
  aiUsage: string;
  apis: string[];
  databaseDesign: string[];
  learningOutcomes: string[];
  resumeBulletPoints: string[];
  interviewQuestions: string[];
  estimatedBuildTime: string;
  monetizationPotential: string;
}

export type GenerateResponseSource = typeof GenerateResponseSource[keyof typeof GenerateResponseSource];


export const GenerateResponseSource = {
  ollama: 'ollama',
  fallback: 'fallback',
} as const;

export interface GenerateResponse {
  source: GenerateResponseSource;
  /** @nullable */
  model: string | null;
  projects: Project[];
  /** @nullable */
  message?: string | null;
}

export type AiStatusProvider = typeof AiStatusProvider[keyof typeof AiStatusProvider];


export const AiStatusProvider = {
  ollama: 'ollama',
  fallback: 'fallback',
} as const;

export interface AiStatus {
  provider: AiStatusProvider;
  available: boolean;
  /** @nullable */
  model: string | null;
  installed: boolean;
  message: string;
}

export interface AiModels {
  models: string[];
  /** @nullable */
  selected: string | null;
  available: boolean;
}