# lib/api-zod/src/generated/types/aiModels.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiModels {
  models: string[];
  /** @nullable */
  selected: string | null;
  available: boolean;
}