# lib/api-zod/src/generated/types/aiStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiStatusProvider } from './aiStatusProvider';

export interface AiStatus {
  provider: AiStatusProvider;
  available: boolean;
  /** @nullable */
  model: string | null;
  installed: boolean;
  message: string;
}