# lib/api-zod/src/generated/types/generateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateInputVariationProject } from './generateInputVariationProject';
import type { Preferences } from './preferences';
import type { Profile } from './profile';

export interface GenerateInput {
  profile: Profile;
  preferences: Preferences;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  variationProject?: GenerateInputVariationProject;
  /** @nullable */
  variationRequest?: string | null;
}