# lib/api-zod/src/generated/types/generateResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateResponseSource } from './generateResponseSource';
import type { Project } from './project';

export interface GenerateResponse {
  source: GenerateResponseSource;
  /** @nullable */
  model: string | null;
  projects: Project[];
  /** @nullable */
  message?: string | null;
}