# lib/api-zod/src/generated/types/preferences.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Preferences {
  category: string;
  difficulty: string;
  /**
     * @minimum 1
     * @maximum 8
     */
  count: number;
  preferredTechnologies?: string[];
}