# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  degree?: string;
  branch?: string;
  year?: string;
  university?: string;
  skills: string[];
  targetRole: string;
  targetCompany?: string;
  careerGoal: string;
  experienceLevel: string;
  existingProjects?: string;
  internshipExperience?: string;
  learningAreas?: string;
}