# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TechStack } from './techStack';

export interface Project {
  id: string;
  title: string;
  oneLinePitch: string;
  problem: string;
  solution: string;
  targetUsers: string[];
  category: string;
  difficulty: string;
  resumeValue: number;
  technicalDepth: number;
  uniqueness: number;
  placementRelevance: number;
  techStack: TechStack;
  coreFeatures: string[];
  advancedFeatures: string[];
  architecture: string;
  aiUsage: string;
  apis: string[];
  databaseDesign: string[];
  learningOutcomes: string[];
  resumeBulletPoints: string[];
  interviewQuestions: string[];
  estimatedBuildTime: string;
  monetizationPotential: string;
}