# lib/api-zod/src/generated/types/techStack.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TechStack {
  frontend: string[];
  backend: string[];
  database: string[];
  ai: string[];
  other: string[];
}